pv_plotter/requirements.txt
asgiref==3.5.2
attrs==22.1.0
click==8.1.3
Django==4.1.1
django-extensions==3.2.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
django-cors-headers==3.11.0
gunicorn==20.1.0
h11==0.13.0
hashids==1.3.1
iniconfig==1.1.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
uvicorn==0.18.3
plotly==5.10.0
pandas==1.5.0
ipython==8.5.0